Move RuntimeCase classes out of YamlSpecTestBase
